put outcomeType in a separate file
