+++ src/DetailsView/components/left-nav/get-status-for-test.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 import { ManualTestStatus } from '../../../common/types/manual-test-status';
-import { OutcomeStats } from '../../reports/components/outcome-type';
+import { OutcomeStats } from '../../reports/components/requirement-outcome-type';
 
 export const getStatusForTest = (stats: OutcomeStats): ManualTestStatus => {
     if (stats.incomplete > 0) {
