+++ src/DetailsView/components/test-steps-nav.tsx
@@ -9,7 +9,7 @@ import { getRequirementsResults } from '../../common/assessment/requirement';
 import { ManualTestStatus, ManualTestStatusData } from '../../common/types/manual-test-status';
 import { VisualizationType } from '../../common/types/visualization-type';
 import { DetailsViewActionMessageCreator } from '../actions/details-view-action-message-creator';
-import { OutcomeTypeSemantic } from '../reports/components/outcome-type';
+import { OutcomeTypeSemantic } from '../reports/components/requirement-outcome-type';
 import { RequirementLink } from './requirement-link';
 
 export interface TestStepNavDeps {
