+++ src/DetailsView/reports/components/outcome-chip-set.tsx
@@ -4,11 +4,11 @@ import * as React from 'react';
 
 import { NamedSFC } from '../../../common/react/named-sfc';
 import { OutcomeChip } from './outcome-chip';
-import { allOutcomeTypes, OutcomeStats } from './outcome-type';
+import { allRequirementOutcomeTypes, OutcomeStats } from './requirement-outcome-type';
 
 export const OutcomeChipSet = NamedSFC<OutcomeStats>('OutcomeChipSet', props => (
     <div className="outcome-chip-set">
-        {allOutcomeTypes.map(outcomeType =>
+        {allRequirementOutcomeTypes.map(outcomeType =>
             props[outcomeType] ? <OutcomeChip key={outcomeType} outcomeType={outcomeType} count={props[outcomeType]} /> : null,
         )}
     </div>
