+++ src/DetailsView/reports/components/outcome-icon-set.tsx
@@ -5,7 +5,7 @@ import * as React from 'react';
 
 import { NamedSFC } from '../../../common/react/named-sfc';
 import { OutcomeIcon } from './outcome-icon';
-import { allOutcomeTypes, OutcomeStats, outcomeTypeSemantics } from './outcome-type';
+import { allRequirementOutcomeTypes, OutcomeStats, outcomeTypeSemantics } from './requirement-outcome-type';
 
 function getText(stats: OutcomeStats): string {
     function textForOutcome(outcomeType): string {
@@ -13,7 +13,7 @@ function getText(stats: OutcomeStats): string {
         const { pastTense } = outcomeTypeSemantics[outcomeType];
         return `${count} ${pastTense}`;
     }
-    return join(allOutcomeTypes.map(textForOutcome), ', ');
+    return join(allRequirementOutcomeTypes.map(textForOutcome), ', ');
 }
 
 export const OutcomeIconSet = NamedSFC<OutcomeStats>('OutcomeIconSet', props => {
@@ -21,7 +21,7 @@ export const OutcomeIconSet = NamedSFC<OutcomeStats>('OutcomeIconSet', props => 
 
     return (
         <div className="outcome-icon-set" title={text}>
-            {allOutcomeTypes.map(outcomeType =>
+            {allRequirementOutcomeTypes.map(outcomeType =>
                 times(props[outcomeType], index => (
                     <span className={'outcome-icon outcome-icon-' + outcomeType} key={`outcome-icon-index-${index}`}>
                         <OutcomeIcon outcomeType={outcomeType} />
