+++ src/DetailsView/reports/components/outcome-summary-bar.tsx
@@ -4,7 +4,7 @@ import * as React from 'react';
 
 import { NamedSFC } from '../../../common/react/named-sfc';
 import { OutcomeIcon } from './outcome-icon';
-import { allOutcomeTypes, OutcomeStats } from './outcome-type';
+import { allRequirementOutcomeTypes, OutcomeStats } from './requirement-outcome-type';
 
 type OutcomeUnits = 'percentage' | 'requirements';
 
@@ -18,7 +18,7 @@ export type OutcomeSummaryBarProps = OutcomeStats & { units?: OutcomeUnits };
 
 export const OutcomeSummaryBar = NamedSFC<OutcomeSummaryBarProps>('OutcomeSummaryBar', props => (
     <div className="outcome-summary-bar">
-        {allOutcomeTypes.map(outcomeType => {
+        {allRequirementOutcomeTypes.map(outcomeType => {
             const { units } = props;
             const count = props[outcomeType];
             const suffix = units === 'percentage' ? '%' : '';
