+++ src/common/assessment/assessment-test-result.ts
@@ -2,7 +2,7 @@
 // Licensed under the MIT License.
 import { AssessmentsProvider } from '../../assessments/types/assessments-provider';
 import { Assessment } from '../../assessments/types/iassessment';
-import { OutcomeStats, outcomeStatsFromManualTestStatus } from '../../DetailsView/reports/components/outcome-type';
+import { OutcomeStats, outcomeStatsFromManualTestStatus } from '../../DetailsView/reports/components/requirement-outcome-type';
 import { ManualTestStatusData } from '../types/manual-test-status';
 import { AssessmentData } from '../types/store-data/assessment-result-data';
 import { VisualizationType } from '../types/visualization-type';
