+++ src/tests/unit/tests/DetailsView/reports/components/outcome-icon.test.tsx
@@ -6,11 +6,11 @@ import { CheckIcon } from '../../../../../../common/icons/check-icon';
 import { CircleIcon } from '../../../../../../common/icons/circle-icon';
 import { CrossIcon } from '../../../../../../common/icons/cross-icon';
 import { OutcomeIcon } from '../../../../../../DetailsView/reports/components/outcome-icon';
-import { allOutcomeTypes } from '../../../../../../DetailsView/reports/components/outcome-type';
+import { allRequirementOutcomeTypes } from '../../../../../../DetailsView/reports/components/requirement-outcome-type';
 
 describe('OutcomeIcon', () => {
     describe('render', () => {
-        allOutcomeTypes.forEach(outcomeType => {
+        allRequirementOutcomeTypes.forEach(outcomeType => {
             test(outcomeType, () => {
                 const wrapper = shallow(<OutcomeIcon outcomeType={outcomeType} />);
                 expect(wrapper.getElement()).toMatchSnapshot();
